Cast modifications

+++ RPG/RPG/barbarian.h
@@ -38,7 +38,7 @@ namespace JD {
 	template<typename T, typename Ability_t, typename Weapon_t>
 	int Barbarian<T, Ability_t, Weapon_t>::getAttackAmnt() const {
 
-		return m_baseAttack + (static_cast<double>(m_weapon[0]) / 2) + (static_cast<double>(m_weapon[1]) / 2);
+		return static_cast<int>(m_baseAttack + (static_cast<double>(m_weapon[0]) / 2) + (static_cast<double>(m_weapon[1]) / 2));
 	}
 
 	template<typename T, typename Ability_t, typename Weapon_t>
